requirements/base.txt
certifi==2021.5.30
charset-normalizer==2.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==2.2.16
django-2gis-maps==1.0.3
django-ckeditor==6.1.0
django-cors-headers==3.8.0
django-debug-toolbar==3.2.2
django-environ==0.4.5
django-js-asset==1.2.2
django-solo==1.1.5
djangorestframework==3.12.4
drf-yasg==1.20.0
idna==3.2
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.1
MarkupSafe==2.0.1
packaging==21.0
Pillow==8.1.0
psycopg2-binary==2.8.5
pyparsing==2.4.7
pytz==2020.5
requests==2.26.0
ruamel.yaml==0.17.13
ruamel.yaml.clib==0.2.6
sqlparse==0.4.1
Unidecode==1.2.0
uritemplate==3.0.1
urllib3==1.26.6
